+++ src/__init__.py
@@ -2,3 +2,4 @@ from config import load_config
 from work_queue import TTWorkQueue
 from callback import callback_addr
 from is_dev_vm import is_dev_vm
+from level_dict import LevelDict