+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## v0.0.10
+* get rid of LevelJsonDict class
+* use Pickle to serialize objects into LevelDB
+
+## v0.0.9
+* added LevelJsonDict class
+
+## v0.0.8
+* added LevelDict class
+
 ## v0.0.7
 * added tornado to install_requires